refactor(코드 구조 개선) : 유효성 검사 폴더를 새로 생성
이전의 InputView 클래스에서는 사용자에게 어떤 화면을 보여주는지 뿐만
아니라 유효성 검사 부분에 대한 코드도 작성되어있었다. 즉, 두가지 일을
하고 있었기에 유효성 검사하는 부분은 다른 클래스로 분리하였다.
그러고나니 좀 더 코드가 읽기 편해졌다.
